feat: implement v0.2 architectural refactor
- Replace double time with long ticks throughout (TickHz = 20)
- Add RoomId to WorldItem and ActorState (default 'beach')
- Move ITickableWorldItem to Abstractions layer
- Update Tick() signature to use long currentTick
- Change TraceEvent.Time to Tick (long), add TimeSeconds helper
- Change Signal.Tick to long
- Change ActionSpec.EstimatedDurationTicks to long
- Change ActionOutcome.ActualDurationTicks to long
- Add deterministic ordering (sort by Id, tie-break by score/actionId)
- Add StalactiteItem in cave room (drips every 60 ticks)
- Delete Office domain (src/JohnnyLike.Domain.Office)
- Delete Office tests (tests/JohnnyLike.Domain.Office.Tests)
- Update IDomainPack interface (double -> long in all relevant methods)
- Update Engine.AdvanceTicks(long), Engine.CurrentTick, Engine.CurrentSeconds
- Update Director, ReservationTable, VarietyMemory to use long ticks
- Update all Island items with RoomId = 'beach'
- Update FakeExecutor.AdvanceTicks(long)
- Add v0.2 tests: determinism, rooms, topo sort, stalactite drip
- Fix TreasureChestItem Failure damage (300L -> 30.0)
- Rewrite scenario tests for Island domain

+++ src/JohnnyLike.Domain.Abstractions/IResourceAvailability.cs
@@ -20,7 +20,7 @@ public interface IResourceAvailability
     /// <param name="utilityId">Unique identifier for the entity reserving (for debugging).</param>
     /// <param name="until">Time until which the resource should be reserved.</param>
     /// <returns>True if reservation succeeded, false if resource is already reserved.</returns>
-    bool TryReserve(ResourceId resourceId, string utilityId, double until);
+    bool TryReserve(ResourceId resourceId, string utilityId, long until);
 
     /// <summary>
     /// Releases a resource reservation.
