+++ tests/JohnnyLike.Domain.Island.Tests/TestHelpers.cs
@@ -10,6 +10,6 @@ namespace JohnnyLike.Domain.Island.Tests;
 public class EmptyResourceAvailability : IResourceAvailability
 {
     public bool IsReserved(ResourceId resourceId) => false;
-    public bool TryReserve(ResourceId resourceId, string utilityId, double until) => true;
+    public bool TryReserve(ResourceId resourceId, string utilityId, long until) => true;
     public void Release(ResourceId resourceId) { }
 }
